day 4g phase a: move ExpertStack to models/heads/
Resolves the lift #1 cross-package leak — the spine's FlowMatchingHead
no longer reaches into reflex.exporters/* for its expert primitive.

Move ExpertStack + ExpertGQALayer + _DecomposedRoPE + _sinusoidal_pos_embedding
from `exporters/smolvla_exporter.py` to `models/heads/expert_stack.py`. Three
existing callsites (pi0_exporter, pi0_prefix_exporter, scripts/local_single_layer_diff)
updated to import from the new location. The OLD location re-exports for
backwards compat — any external code that still imports from
`reflex.exporters.smolvla_exporter` keeps working.

Phase B (next PR) wires Pi0VLA.predict_action + the parity gate vs the
legacy pi0_exporter path. Phase A here is the architectural prerequisite
— gives the spine a clean home for the expert primitive before adding
the inference orchestration on top.

Test plan:
- 2602 tests passed locally (pi0_vla, vlm_prefix, flow_matching_head all green)
- Static check: is-comparison confirms re-export to canonical class identity
