+++ scripts/local_single_layer_diff.py
@@ -14,7 +14,7 @@ import torch.nn.functional as F
 
 def main():
     from lerobot.policies.smolvla.modeling_smolvla import SmolVLAPolicy
-    from reflex.exporters.smolvla_exporter import ExpertGQALayer, _DecomposedRoPE
+    from reflex.models.heads.expert_stack import ExpertGQALayer, _DecomposedRoPE
 
     policy = SmolVLAPolicy.from_pretrained("lerobot/smolvla_libero")
     policy.eval().to(dtype=torch.float32).to("cpu")
